+++ app/src/main/java/org/simple/clinic/login/pin/LoginPinScreenController.kt
@@ -13,7 +13,7 @@ import org.simple.clinic.user.UserSession
 import org.simple.clinic.widgets.UiEvent
 import javax.inject.Inject
 
-typealias Ui = LoginPinScreen
+typealias Ui = LoginPinScreenUi
 typealias UiChange = (Ui) -> Unit
 
 class LoginPinScreenController @Inject constructor(
@@ -23,7 +23,6 @@ class LoginPinScreenController @Inject constructor(
   override fun apply(events: Observable<UiEvent>): ObservableSource<UiChange> {
     val replayedEvents = ReplayUntilScreenIsDestroyed(events)
         .compose(UpdateLoginEntryWithEnteredPin(userSession))
-        .compose(ReportAnalyticsEvents())
         .replay()
 
     return Observable.mergeArray(