+++ app/src/main/java/org/simple/clinic/login/pin/UiActions.kt
@@ -1,0 +1,4 @@
+package org.simple.clinic.login.pin
+
+interface UiActions {
+}